chore(executor): extract row_to_json.rs (#116 step 7) (#517)
Pulls the schema-driven dynamic row-to-JSON decoders out of mod.rs:

- row_to_json (PG) — column-by-column FieldType → JSON.
- row_to_json_my (MySQL) — same shape over MySqlRow.
- row_to_json_sqlite (SQLite) — handles SQLite TEXT-affinity quirks
  (Date/DateTime stored as TEXT, JSON round-tripped through string).
- hex_encode helper for FieldType::Binary.
- augment_joined_columns_{pg,my,sqlite} — adds joined-table
  `<alias>__<col>` columns to the JSON output.
- select_rows_as_json / select_one_row_as_json — tri-dialect Pool
  dispatchers that wire compile_select → bind_query family →
  row_to_json family.

These are consumed by the admin/viewset list-retrieve handlers
and contenttypes::fetch_row_as_json (#80, #89). Pulled into their
own file because the three per-dialect impls add up to 500 LOC of
near-mechanical scaffolding.

mod.rs: 3412 → 2923 LOC (-489). row_to_json.rs: 524 LOC.

1482 lib tests pass on the sqlite,tenancy litmus.
